Computer Vision/Handwritten Note to Markdown/Source Code/requirements.txt
transformers
torch
pillow
opencv-python
numpy
pyyaml
